pyproject.toml
[tool.poetry]
name = "mastodon-bdsm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mastodon_bdsm"}]

[tool.poetry.dependencies]
python = "3.11.3"
wordcloud = "^1.9.1.1"
blurhash = "1.1.4"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
click = "8.1.3"
contourpy = "1.0.6"
cycler = "0.11.0"
decorator = "5.1.1"
flask = "2.2.2"
flask-sqlalchemy = "3.0.2"
fonttools = "4.38.0"
greenlet = "2.0.1"
idna = "3.4"
importlib-metadata = "5.0.0"
itsdangerous = "2.1.2"
jieba = "0.42.1"
jinja2 = "3.1.2"
kiwisolver = "1.4.4"
markupsafe = "2.1.1"
mastodon-py = "1.8.0"
matplotlib = "3.6.2"
numpy = "1.24.1"
packaging = "22.0"
pillow = "9.3.0"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
python-magic = "0.4.27"
pytz = "2022.6"
requests = "2.28.1"
six = "1.16.0"
sqlalchemy = "1.4.45"
tenacity = "8.1.0"
typing-extensions = "4.4.0"
urllib3 = "1.26.13"
werkzeug = "2.2.2"
zipp = "3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
